backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic[email]
bcrypt
python-jose[cryptography]
python-dotenv
